Attempt to lessen memory usage of call_spots_pixel
Also, now detects if output files have already been made.

+++ @iss/get_spot_colors.m
@@ -36,5 +36,5 @@ end
 %Only keep spots that are in all rounds/channels
 SpotColorsToUse = TileSpotColors(:,o.UseChannels,o.UseRounds);
 Good = all(isfinite(SpotColorsToUse(:,:)),2);
-GoodAnchorLocalYX = AnchorLocalYX(Good,1:2);
-GoodSpotColors = TileSpotColors(Good,:,:);
+GoodAnchorLocalYX = int32(AnchorLocalYX(Good,1:2));
+GoodSpotColors = int32(TileSpotColors(Good,:,:));
